requirements.txt
requests
bs4
pdfminer.six
jieba
wordcloud
Pillow
matplotlib